refactor legal code
