+++ tests/test_legal/test_rm_load.py
@@ -55,9 +55,9 @@ class TestGetRulemaking(BaseTestCase):
             }
         )
 
-    # @patch("webservices.legal_docs.advisory_opinions.get_bucket")
-    # @patch("webservices.legal_docs.advisory_opinions.create_es_client")
-    # @patch("webservices.legal_docs.es_management.create_index")
+    # @patch("webservices.legal.legal_docs.advisory_opinions.get_bucket")
+    # @patch("webservices.legal.legal_docs.advisory_opinions.create_es_client")
+    # @patch("webservices.legal.legal_docs.es_management.create_index")
     # def test_ao_with_entities(self, get_bucket, create_es_client, create_index):
     #     expected_requestor_names = [
     #         "The Manchurian Candidate",
@@ -105,9 +105,9 @@ class TestGetRulemaking(BaseTestCase):
     #         expected_representative_names
     #     )
 
-    # @patch("webservices.legal_docs.advisory_opinions.get_bucket")
-    # @patch("webservices.legal_docs.advisory_opinions.create_es_client")
-    # @patch("webservices.legal_docs.es_management.create_index")
+    # @patch("webservices.legal.legal_docs.advisory_opinions.get_bucket")
+    # @patch("webservices.legal.legal_docs.advisory_opinions.create_es_client")
+    # @patch("webservices.legal.legal_docs.es_management.create_index")
     # def test_ao_with_entity_individual(self, get_bucket, create_es_client, create_index):
     #     expected_entity = {
     #         "role": "Commenter",
@@ -136,9 +136,9 @@ class TestGetRulemaking(BaseTestCase):
     #             "name": "Mr Dan Becker MD",
     #             "type": "Individual"}]
 
-    # @patch("webservices.legal_docs.advisory_opinions.get_bucket")
-    # @patch("webservices.legal_docs.advisory_opinions.create_es_client")
-    # @patch("webservices.legal_docs.es_management.create_index")
+    # @patch("webservices.legal.legal_docs.advisory_opinions.get_bucket")
+    # @patch("webservices.legal.legal_docs.advisory_opinions.create_es_client")
+    # @patch("webservices.legal.legal_docs.es_management.create_index")
     # def test_completed_ao_with_docs(self, get_bucket, create_es_client, create_index):
     #     ao_no = "2017-01"
     #     filename = "Some File.pdf"
@@ -173,9 +173,9 @@ class TestGetRulemaking(BaseTestCase):
     #     for key in expected_document:
     #         assert actual_document[key] == expected_document[key]
 
-    # @patch("webservices.legal_docs.advisory_opinions.get_bucket")
-    # @patch("webservices.legal_docs.advisory_opinions.create_es_client")
-    # @patch("webservices.legal_docs.es_management.create_index")
+    # @patch("webservices.legal.legal_docs.advisory_opinions.get_bucket")
+    # @patch("webservices.legal.legal_docs.advisory_opinions.create_es_client")
+    # @patch("webservices.legal.legal_docs.es_management.create_index")
     # def test_ao_citations(self, get_bucket, create_es_client, create_index):
     #     ao1_document = {
     #         "document_id": 1,
@@ -230,9 +230,9 @@ class TestGetRulemaking(BaseTestCase):
     #     assert actual_ao2["ao_citations"] == [{"no": "2017-01", "name": "1st AO name"}]
     #     assert actual_ao2["aos_cited_by"] == []
 
-    # @patch("webservices.legal_docs.advisory_opinions.get_bucket")
-    # @patch("webservices.legal_docs.advisory_opinions.create_es_client")
-    # @patch("webservices.legal_docs.es_management.create_index")
+    # @patch("webservices.legal.legal_docs.advisory_opinions.get_bucket")
+    # @patch("webservices.legal.legal_docs.advisory_opinions.create_es_client")
+    # @patch("webservices.legal.legal_docs.es_management.create_index")
     # def test_ao_offsets(self, get_bucket, create_es_client, create_index):
     #     expected_ao1 = {
     #         "type": "advisory_opinions",
