+++ webservices/legal/legal_docs/responses.py
