+++ webservices/legal/utils_es.py
@@ -1,507 +1,184 @@
 import logging
-import datetime
 import json
 import time
+import datetime
+import certifi
+import webservices.legal.constants as constants
 
-from webservices.utils import (
-    create_es_client,
-    get_service_instance,
-    get_service_instance_credentials,
-    DateTimeEncoder,
-)
+
+from webservices.legal.mappings import rm_mapping, ao_mapping, arch_mur_mapping, case_mapping
 from webservices.env import env
 from webservices.tasks.utils import get_bucket
+
+from elasticsearch import Elasticsearch, RequestsHttpConnection
+from requests_aws4auth import AWS4Auth
+from json import JSONEncoder
 
 logger = logging.getLogger(__name__)
 # To debug, uncomment the line below:
 # logger.setLevel(logging.DEBUG)
 
-CASE_INDEX = "case_index"
-CASE_ALIAS = "case_alias"
-CASE_SWAP_INDEX = "case_swap_index"
-CASE_REPO = "case_repo"
-CASE_SNAPSHOT = "case_snapshot"
-
-AO_INDEX = "ao_index"
-AO_ALIAS = "ao_alias"
-AO_SWAP_INDEX = "ao_swap_index"
-AO_REPO = "ao_repo"
-AO_SNAPSHOT = "ao_snapshot"
-
-ARCH_MUR_INDEX = "arch_mur_index"
-ARCH_MUR_ALIAS = "arch_mur_alias"
-ARCH_MUR_SWAP_INDEX = "arch_mur_swap_index"
-ARCH_MUR_REPO = "arch_mur_repo"
-ARCH_MUR_SNAPSHOT = "arch_mur_snapshot"
-
-SEARCH_ALIAS = "search_alias"
-S3_BACKUP_DIRECTORY = "es-backups"
-S3_PRIVATE_SERVICE_INSTANCE_NAME = "fec-s3-snapshot"
-
-TEST_CASE_INDEX = "test_case_index"
-TEST_CASE_ALIAS = "test_case_alias"
-TEST_AO_INDEX = "test_ao_index"
-TEST_AO_ALIAS = "test_ao_alias"
-TEST_ARCH_MUR_INDEX = "test_arch_mur_index"
-TEST_ARCH_MUR_ALIAS = "test_arch_mur_alias"
-TEST_SEARCH_ALIAS = "test_search_alias"
-
-DOCS_PATH = "docs"
-
-SORT_MAPPING = {
-    "sort1": {"type": "integer"},
-    "sort2": {"type": "integer"},
-}
-
-# ==== start define XXXX_MAPPING for index: XXXX_INDEX
-CASE_DOCUMENT_MAPPING = {
-    "type": "nested",
-    "properties": {
-        "document_id": {"type": "long"},
-        "category": {"type": "keyword"},
-        "description": {"type": "text"},
-        "length": {"type": "long"},
-        "text": {
-            "type": "text",
-            "term_vector": "with_positions_offsets",
-        },
-        "document_date": {"type": "date", "format": "dateOptionalTime"},
-        "url": {"type": "text"},
-        "filename": {"type": "keyword"},
-        "doc_order_id": {"type": "integer"},
-    },
-}
-
-ADMIN_FINE_MAPPING = {
-    "type": {"type": "keyword"},
-    "doc_id": {"type": "keyword"},
-    "no": {"type": "keyword"},
-    "case_serial": {"type": "integer"},
-    "name": {"type": "text"},
-    "published_flg": {"type": "boolean"},
-    "commission_votes": {
-        "properties": {
-            "vote_date": {"type": "date", "format": "dateOptionalTime"},
-            "action": {"type": "text"},
-        }
-    },
-    "documents": CASE_DOCUMENT_MAPPING,
-    "url": {"type": "text", "index": False},
-    "committee_id": {"type": "text"},
-    "report_year": {"type": "keyword"},
-    "report_type": {"type": "text", "index": False},
-    "reason_to_believe_action_date": {
-        "type": "date",
-        "format": "dateOptionalTime",
-    },
-    "reason_to_believe_fine_amount": {"type": "long"},
-    "challenge_receipt_date": {
-        "type": "date",
-        "format": "dateOptionalTime",
-    },
-    "challenge_outcome": {"type": "text", "index": False},
-    "final_determination_date": {
-        "type": "date",
-        "format": "dateOptionalTime",
-    },
-    "final_determination_amount": {"type": "long"},
-    "payment_amount": {"type": "long", "index": False},
-    "treasury_referral_date": {
-        "type": "date",
-        "format": "dateOptionalTime",
-    },
-    "treasury_referral_amount": {"type": "long", "index": False},
-    "petition_court_filing_date": {
-        "type": "date",
-        "format": "dateOptionalTime",
-    },
-    "petition_court_decision_date": {
-        "type": "date",
-        "format": "dateOptionalTime",
-    },
-    "civil_penalty_payment_status": {
-        "type": "text",
-        "index": False,
-    },
-    "civil_penalty_due_date": {
-        "type": "date",
-        "format": "dateOptionalTime",
-    },
-    "dispositions": {
-        "properties": {
-            "disposition_description": {
-                "type": "text",
-                "index": False,
-            },
-            "disposition_date": {
-                "type": "date",
-                "format": "dateOptionalTime",
-            },
-            "penalty": {
-                "type": "scaled_float",
-                "scaling_factor": 100,
-                "index": False,
-                "null_value": -1
-            },
-        }
-    },
-}
-
-CITATION_MAPPING = {
-    "type": {"type": "keyword"},
-    "citation_type": {"type": "keyword"},
-    "citation_text": {"type": "keyword"},
-    "doc_type": {"type": "keyword"},
-}
-
-MUR_MAPPING = {
-    "type": {"type": "keyword"},
-    "doc_id": {"type": "keyword"},
-    "no": {"type": "keyword"},
-    "case_serial": {"type": "integer"},
-    "name": {"type": "text"},
-    "published_flg": {"type": "boolean"},
-    "commission_votes": {
-        "properties": {
-            "vote_date": {"type": "date", "format": "dateOptionalTime"},
-            "action": {"type": "text"},
-        }
-    },
-    "documents": CASE_DOCUMENT_MAPPING,
-    "url": {"type": "text", "index": False},
-    "mur_type": {"type": "keyword"},
-    "subjects": {
-        "type": "nested",
-        "properties": {
-            "subject": {"type": "text"},
-            "primary_subject_id": {"type": "keyword"},
-            "secondary_subject_id": {"type": "keyword"},
-        }
-    },
-    "election_cycles": {"type": "long"},
-    "participants": {
-        "properties": {
-            "citations": {"type": "object"},
-            "name": {"type": "text"},
-            "role": {"type": "text"},
-        }
-    },
-    "respondents": {"type": "text"},
-    "dispositions": {
-        "type": "nested",
-        "properties": {
-            "citations": {
-                "type": "nested",
-                "properties": {
-                    "text": {"type": "text"},
-                    "title": {"type": "text"},
-                    "type": {"type": "text"},
-                    "url": {"type": "text"},
-                }
-            },
-            "disposition": {"type": "text"},
-            "penalty": {"type": "scaled_float", "scaling_factor": 100, "null_value": -1},
-            "respondent": {"type": "text"},
-            "mur_disposition_catagory_id": {"type": "keyword"},
-        }
-    },
-    "open_date": {"type": "date", "format": "dateOptionalTime"},
-    "close_date": {"type": "date", "format": "dateOptionalTime"},
-}
-
-ADR_MAPPING = {
-    "type": {"type": "keyword"},
-    "doc_id": {"type": "keyword"},
-    "no": {"type": "keyword"},
-    "case_serial": {"type": "integer"},
-    "name": {"type": "text"},
-    "published_flg": {"type": "boolean"},
-    "complainant": {"type": "text"},
-    "commission_votes": {
-        "properties": {
-            "vote_date": {"type": "date", "format": "dateOptionalTime"},
-            "action": {"type": "text"},
-            "commissioner_name": {"type": "text"},
-            "vote_type": {"type": "text"},
-        }
-    },
-    "non_monetary_terms": {"type": "text"},
-    "non_monetary_terms_respondents": {"type": "text"},
-    "documents": CASE_DOCUMENT_MAPPING,
-    "url": {"type": "text", "index": False},
-    "mur_type": {"type": "keyword"},
-    "subjects": {
-        "type": "nested",
-        "properties": {
-            "subject": {"type": "text"},
-            "primary_subject_id": {"type": "keyword"},
-            "secondary_subject_id": {"type": "keyword"},
-        }
-    },
-    "election_cycles": {"type": "long"},
-    "participants": {
-        "properties": {
-            "name": {"type": "text"},
-            "role": {"type": "text"},
-        }
-    },
-    "respondents": {"type": "text"},
-    "case_status": {"type": "text"},
-    "dispositions": {
-        "type": "nested",
-        "properties": {
-            "citations": {
-                "type": "nested",
-                "properties": {
-                    "text": {"type": "text"},
-                    "title": {"type": "text"},
-                    "type": {"type": "text"},
-                    "url": {"type": "text"},
-                }
-            },
-            "disposition": {"type": "text"},
-            "penalty": {"type": "scaled_float", "scaling_factor": 100, "null_value": -1},
-            "respondent": {"type": "text"},
-        }
-    },
-    "open_date": {"type": "date", "format": "dateOptionalTime"},
-    "close_date": {"type": "date", "format": "dateOptionalTime"},
-}
-
-STATUTE_MAPPING = {
-    "type": {"type": "keyword"},
-    "doc_id": {"type": "keyword"},
-    "name": {"type": "text"},
-    "text": {"type": "text"},
-    "no": {"type": "keyword"},
-    "title": {"type": "text"},
-    "chapter": {"type": "text"},
-    "subchapter": {"type": "text"},
-    "url": {"type": "text", "index": False},
-}
-
-ALL_MAPPING = {}
-ALL_MAPPING.update(ADMIN_FINE_MAPPING)
-ALL_MAPPING.update(CITATION_MAPPING)
-ALL_MAPPING.update(ADR_MAPPING)
-ALL_MAPPING.update(MUR_MAPPING)
-ALL_MAPPING.update(STATUTE_MAPPING)
-ALL_MAPPING.update(SORT_MAPPING)
-
-CASE_MAPPING = {"properties": ALL_MAPPING}
-# ==== end define CASE_MAPPING for index: CASE_INDEX
-
-# ==== start define AO_MAPPING for index: AO_INDEX
-AO_MAPPING = {
-    "dynamic": "false",
-    "properties": {
-        "type": {"type": "keyword"},
-        "no": {"type": "keyword"},
-        "ao_no": {"type": "keyword"},
-        "ao_serial": {"type": "integer"},
-        "ao_year": {"type": "integer"},
-        "doc_id": {"type": "keyword"},
-        "name": {"type": "text"},
-        "summary": {"type": "text"},
-        "request_date": {"type": "date", "format": "dateOptionalTime"},
-        "issue_date": {"type": "date", "format": "dateOptionalTime"},
-        "is_pending": {"type": "boolean"},
-        "status": {"type": "text"},
-        "published_flg": {"type": "boolean"},
-        "ao_citations": {
-            "properties": {
-                "name": {"type": "text"},
-                "no": {"type": "text"},
-            }
-        },
-        "aos_cited_by": {
-            "properties": {
-                "name": {"type": "text"},
-                "no": {"type": "text"},
-            }
-        },
-        "statutory_citations": {
-            "type": "nested",
-            "properties": {
-                "title": {"type": "long"},
-                "section": {"type": "text"},
-            },
-        },
-        "regulatory_citations": {
-            "type": "nested",
-            "properties": {
-                "part": {"type": "long"},
-                "title": {"type": "long"},
-                "section": {"type": "long"},
-            },
-        },
-        "documents": {
-            "type": "nested",
-            "properties": {
-                "document_id": {"type": "long"},
-                "ao_doc_category_id": {"type": "keyword"},
-                "category": {"type": "keyword"},
-                "description": {"type": "text"},
-                "text": {
-                    "type": "text",
-                    "term_vector": "with_positions_offsets",
-                },
-                "date": {"type": "date", "format": "dateOptionalTime"},
-                "url": {"type": "text", "index": False},
-                "filename": {"type": "keyword"},
-            },
-        },
-        "requestor_names": {"type": "text"},
-        "requestor_types": {"type": "keyword"},
-        "commenter_names": {"type": "text"},
-        "representative_names": {"type": "text"},
-        "citation_type": {"type": "keyword"},
-        "citation_text": {"type": "keyword"},
-        "doc_type": {"type": "keyword"},
-        "entities": {
-            "properties": {
-                "role": {"type": "keyword"},
-                "name": {"type": "text"},
-                "type": {"type": "text"},
-            },
-        },
-        "sort1": {"type": "integer"},
-        "sort2": {"type": "integer"},
-    }
-}
-# ==== end define AO_MAPPING for index: AO_INDEX
-
-# ==== start define ARCH_MUR_MAPPING for index: ARCH_MUR_INDEX
-ARCH_MUR_DOCUMENT_MAPPING = {
-    "type": "nested",
-    "properties": {
-        "document_id": {"type": "integer"},
-        "length": {"type": "long"},
-        "text": {
-            "type": "text",
-            "term_vector": "with_positions_offsets",
-        },
-        "url": {"type": "text", "index": False},
-        "filename": {"type": "keyword"}
-    },
-}
-
-ARCH_MUR_SUBJECT_MAPPING = {
-    "properties": {
-        "text": {"type": "text"},
-        "children": {
-            "properties": {
-                "text": {"type": "text"},
-                "children": {
-                    "properties": {
-                        "text": {"type": "text"}
-                    }
-                }
-            }
-        }
-    }
-}
-
-ARCH_MUR_CITATION_MAPPING = {
-    "properties": {
-        "us_code": {
-            "properties": {"text": {"type": "text"}, "url": {"type": "text"}}
-        },
-        "regulations": {
-            "properties": {"text": {"type": "text"}, "url": {"type": "text"}}
-        }
-    }
-}
-
-ARCH_MUR_MAPPING = {
-    "dynamic": "false",
-    "properties": {
-        "type": {"type": "keyword"},
-        "doc_id": {"type": "keyword"},
-        "no": {"type": "keyword"},
-        "case_serial": {"type": "integer"},
-        "mur_name": {"type": "text"},
-        "mur_type": {"type": "keyword"},
-        "open_date": {"type": "date", "format": "dateOptionalTime"},
-        "close_date": {"type": "date", "format": "dateOptionalTime"},
-        "url": {"type": "text", "index": False},
-        "complainants": {"type": "text"},
-        "respondents": {"type": "text"},
-        "documents": ARCH_MUR_DOCUMENT_MAPPING,
-        "citations": ARCH_MUR_CITATION_MAPPING,
-        "subject": ARCH_MUR_SUBJECT_MAPPING,
-        "sort1": {"type": "integer"},
-        "sort2": {"type": "integer"},
-    }
-}
-# ==== end define ARCH_MUR_MAPPING for index: ARCH_MUR_INDEX
-
-ANALYZER_SETTING = {
-    "analysis": {
-        "analyzer": {
-            "default": {
-                "type": "standard"
-            }
-        }
-    },
-    "highlight.max_analyzed_offset": 60000000,
-}
-
 # SEARCH_ALIAS is used for legal/search endpoint
+# RM_SEARCH_ALIAS is used for rulemaking/search endpoint
 # XXXX_ALIAS is used for load data to XXXX_INDEX on Elasticsearch
 INDEX_DICT = {
-    CASE_INDEX: (CASE_MAPPING, CASE_ALIAS, SEARCH_ALIAS, CASE_SWAP_INDEX,
-                 CASE_REPO, CASE_SNAPSHOT),
-    AO_INDEX: (AO_MAPPING, AO_ALIAS, SEARCH_ALIAS, AO_SWAP_INDEX,
-               AO_REPO, AO_SNAPSHOT),
-    ARCH_MUR_INDEX: (ARCH_MUR_MAPPING, ARCH_MUR_ALIAS, SEARCH_ALIAS, ARCH_MUR_SWAP_INDEX,
-                     ARCH_MUR_REPO, ARCH_MUR_SNAPSHOT),
-    CASE_SWAP_INDEX: (CASE_MAPPING, "", "", "", "", ""),
-    AO_SWAP_INDEX: (AO_MAPPING, "", "", "", "", ""),
-    ARCH_MUR_SWAP_INDEX: (ARCH_MUR_MAPPING, "", "", "", "", ""),
+    constants.CASE_INDEX: (case_mapping.CASE_MAPPING, constants.CASE_ALIAS,
+                           constants.SEARCH_ALIAS, constants.CASE_SWAP_INDEX,
+                           constants.CASE_REPO, constants.CASE_SNAPSHOT),
+    constants.AO_INDEX: (ao_mapping.AO_MAPPING, constants.AO_ALIAS,
+                         constants.SEARCH_ALIAS, constants.AO_SWAP_INDEX,
+                         constants.AO_REPO, constants.AO_SNAPSHOT),
+    constants.ARCH_MUR_INDEX: (arch_mur_mapping.ARCH_MUR_MAPPING, constants.ARCH_MUR_ALIAS, constants.SEARCH_ALIAS,
+                               constants.ARCH_MUR_SWAP_INDEX,
+                               constants.ARCH_MUR_REPO, constants.ARCH_MUR_SNAPSHOT),
+    constants.CASE_SWAP_INDEX: (case_mapping.CASE_MAPPING, "", "", "", "", ""),
+    constants.AO_SWAP_INDEX: (ao_mapping.AO_MAPPING, "", "", "", "", ""),
+    constants.ARCH_MUR_SWAP_INDEX: (arch_mur_mapping.ARCH_MUR_MAPPING, "", "", "", "", ""),
+    constants.RM_INDEX: (rm_mapping.RM_MAPPING, constants.RM_ALIAS,
+                         constants.RM_SEARCH_ALIAS, constants.RM_SWAP_INDEX,
+                         constants.RM_REPO, constants.RM_SNAPSHOT),
 }
 
 TEST_INDEX_DICT = {
-        TEST_CASE_INDEX: (CASE_MAPPING, TEST_CASE_ALIAS, TEST_SEARCH_ALIAS),
-        TEST_AO_INDEX: (AO_MAPPING, TEST_AO_ALIAS, TEST_SEARCH_ALIAS),
-        TEST_ARCH_MUR_INDEX: (ARCH_MUR_MAPPING, TEST_ARCH_MUR_ALIAS, TEST_SEARCH_ALIAS)
-    }
+        constants.TEST_CASE_INDEX: (case_mapping.CASE_MAPPING,
+                                    constants.TEST_CASE_ALIAS, constants.TEST_SEARCH_ALIAS),
+        constants.TEST_AO_INDEX: (ao_mapping.AO_MAPPING, constants.TEST_AO_ALIAS,
+                                  constants.TEST_SEARCH_ALIAS),
+        constants.TEST_ARCH_MUR_INDEX: (arch_mur_mapping.ARCH_MUR_MAPPING,
+                                        constants.TEST_ARCH_MUR_ALIAS, constants.TEST_SEARCH_ALIAS)
+}
+
+
+def get_service_instance(service_instance_name):
+    return env.get_service(name=service_instance_name)
+
+
+def get_service_instance_credentials(service_instance):
+    return service_instance.credentials
+
+
+def create_eregs_link(part, section):
+    url_part_section = part
+    if section:
+        url_part_section += "-" + section
+    return "/regulations/{}/CURRENT".format(url_part_section)
+
+
+def upload_citations(statutory_citations, regulatory_citations, index, doc_type, es_client):
+    try:
+        for citation in statutory_citations:
+            entry = {
+                "type": "citations",
+                "citation_text": citation,
+                "citation_type": "statute",
+                "doc_type": doc_type,
+            }
+            es_client.index(index, entry, id=doc_type + "_" + citation)
+
+        for citation in regulatory_citations:
+            entry = {
+                "type": "citations",
+                "citation_text": citation,
+                "citation_type": "regulation",
+                "doc_type": doc_type,
+            }
+            es_client.index(index, entry, id=doc_type + "_" + citation)
+    except Exception:
+        logger.error("An error occurred while uploading {} citations".format(doc_type))
+
+
+def check_filter_exists(kwargs, filter):
+    if kwargs.get(filter):
+        for val in kwargs.get(filter):
+            if len(val) > 0:
+                return True
+        return False
+    else:
+        return False
+
+
+# To display the open_date and close_date of JSON format inside object "mur"
+class DateTimeEncoder(JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, (datetime.date, datetime.datetime)):
+            return obj.isoformat()
+
+
+def create_es_client():
+    try:
+        es_service = get_service_instance(constants.ES_SERVICE_INSTANCE_NAME)
+        if es_service:
+            credentials = es_service.credentials
+            #  create "http_auth".
+            host = credentials["host"]
+            access_key = credentials["access_key"]
+            secret_key = credentials["secret_key"]
+            aws_auth = AWS4Auth(access_key, secret_key, constants.REGION, constants.AWS_ES_SERVICE)
+
+            #  create elasticsearch client through "http_auth"
+            es_client = Elasticsearch(
+                hosts=[{"host": host, "port": constants.PORT}],
+                http_auth=aws_auth,
+                use_ssl=True,
+                verify_certs=True,
+                ca_certs=certifi.where(),
+                connection_class=RequestsHttpConnection,
+                timeout=40,
+                max_retries=10,
+                retry_on_timeout=True,
+            )
+        else:
+            # create local elasticsearch client
+            url = "http://localhost:9200"
+            es_client = Elasticsearch(
+                url,
+                timeout=30,
+                max_retries=10,
+                retry_on_timeout=True,
+            )
+        return es_client
+    except Exception as err:
+        logger.error("An error occurred trying to create Elasticsearch client.{0}".format(err))
 
 
 def create_index(index_name=None):
     """
-    Creating an index for storing legal data on Elasticsearch based on 'INDEX_DICT'.
+    Creating an index for storing legal or rulemaking data on Elasticsearch based on 'INDEX_DICT'.
     - 'INDEX_DICT' description:
     1) CASE_INDEX includes DOCUMENT_TYPE=('statutes','murs','adrs','admin_fines')
     'murs' means current mur only.
     2) AO_INDEX includes DOCUMENT_TYPE=('advisory_opinions')
     3) ARCH_MUR_INDEX includes DOCUMENT_TYPE=('murs'), archived mur only
+    4) RM_INDEX includes DOCUMENT_TYPE=('rulemaking')?
 
-    -Two aliases will be created under each index: XXXX_ALIAS and SEARCH_ALIAS
+    -Two aliases will be created under each index: XXXX_ALIAS and SEARCH_ALIAS or RM_SEARCH_ALIAS
     a) XXXX_ALIAS is used for load data to XXXX_INDEX on Elasticsearch
     b) SEARCH_ALIAS is used for '/legal/search/' endpoint
+    c) RM_SEARCH_ALIAS is used for '/rulemaking/search/' endpoint
 
     -How to call this function in python code:
     a) create_index(CASE_INDEX)
     b) create_index(AO_INDEX)
     c) create_index(ARCH_MUR_INDEX)
+    d) create_index(RM_INDEX)
 
     -How to run command from terminal:
-    a) python cli.py create_index case_index (or 'python cli.py create_index')
+    a) python cli.py create_index case_index
     b) python cli.py create_index ao_index
     c) python cli.py create_index arch_mur_index
+    d) python cli.py create_index rm_index
 
     -How to call task command:
     a) cf run-task api --command "python cli.py create_index" -m 2G --name create_case_index
     b) cf run-task api --command "python cli.py create_index ao_index" -m 2G --name create_ao_index
-    b) cf run-task api --command "python cli.py create_index arch_mur_index" -m 2G --name create_arch_mur_index
+    c) cf run-task api --command "python cli.py create_index arch_mur_index" -m 2G --name create_arch_mur_index
+    d) cf run-task api --command "python cli.py create_index rm_index" -m 2G --name create_rm_index
 
     -This function won't allow to create any other index that is not in 'INDEX_DICT'.
     """
-    index_name = index_name or CASE_INDEX
     body = {}
     aliases = {}
-    body.update({"settings": ANALYZER_SETTING})
+    body.update({"settings": constants.ANALYZER_SETTING})
 
     if index_name in INDEX_DICT.keys():
         # Before creating index, delete this index and corresponding aliases first.
@@ -531,20 +208,47 @@ def create_index(index_name=None):
         logger.error(" Invalid index '{0}'.".format(index_name))
 
 
-def create_test_indices():
+def delete_index(index_name):
+    """
+    Delete an index, the argument 'index_name' is required
+    -How to call this function in python code:
+    a) es_delete_index(CASE_INDEX)
+    b) es_delete_index(AO_INDEX)
+    c) es_delete_index(ARCH_MUR_INDEX)
+    d) es_delete_index(RM_INDEX)
 
-    INDEX_DICT.update(TEST_INDEX_DICT)
+    -How to run from terminal:
+    a) python cli.py es_delete_index case_index
+    b) python cli.py es_delete_index ao_index
+    c) python cli.py es_delete_index arch_mur_index
+    d) python cli.py es_delete_index rm_index
 
-    for index_name in TEST_INDEX_DICT:
-        create_index(index_name)
-        INDEX_DICT.pop(index_name)
+    -How to call task command in cf:
+    a) cf run-task api --command "python cli.py delete_index case_index" -m 2G --name delete_case_index
+    b) cf run-task api --command "python cli.py delete_index ao_index" -m 2G --name delete_ao_index
+    c) cf run-task api --command "python cli.py delete_index arch_mur_index" -m 2G --name delete_arch_mur_index
+    d) cf run-task api --command "python cli.py delete_index rm_index" -m 2G --name delete_rm_index
+    """
+    es_client = create_es_client()
+    logger.info(" Checking if index '{0}' already exist...".format(index_name))
+    if es_client.indices.exists(index=index_name):
+        try:
+            logger.info(" Deleting index '{0}'...".format(index_name))
+            es_client.indices.delete(index_name)
+            # sleep 60 seconds (1 min)
+            time.sleep(60)
+            logger.info(" The index '{0}' is deleted successfully.".format(index_name))
+        except Exception:
+            pass
+    else:
+        logger.error(" The index '{0}' is not found.".format(index_name))
 
 
 def display_index_alias():
     """
     Display all indices and aliases on Elasticsearch.
     -How to run from terminal:
-    'python cli.py display_index_alias'
+    'python cli.py es_display_index_alias'
 
     -How to call task command:
     cf run-task api --command "python cli.py display_index_alias" -m 2G --name display_index_alias
@@ -563,54 +267,31 @@ def display_mapping(index_name=None):
     """
     Display the index mapping.
     -How to run from terminal:
-    a) python cli.py display_mapping case_index (or 'python cli.py display_mapping')
-    b) python cli.py display_mapping ao_index
-    c) python cli.py display_mapping arch_mur_index
+    a) python cli.py es_display_mapping case_index
+    b) python cli.py es_display_mapping ao_index
+    c) python cli.py es_display_mapping arch_mur_index
+    d) python cli.py es_display_mapping rm_index
 
     -How to call task command:
     a) cf run-task api --command "python cli.py display_mapping case_index" -m 2G --name display_case_index_mapping
     b) cf run-task api --command "python cli.py display_mapping ao_index" -m 2G --name display_ao_index_mapping
     c) cf run-task api --command "python cli.py display_mapping arch_mur_index" -m 2G
     --name display_arch_mur_index_mapping
+    d) cf run-task api --command "python cli.py display_mapping rm_index" -m 2G --name display_rm_index_mapping
     """
-    index_name = index_name or CASE_INDEX
     es_client = create_es_client()
     logger.info(" The mapping for index '{0}': \n{1}".format(
         index_name,
         json.dumps(es_client.indices.get_mapping(index=index_name), indent=3)))
 
 
-def delete_index(index_name):
-    """
-    Delete an index, the argument 'index_name' is required
-    -How to call this function in python code:
-    a) delete_index(CASE_INDEX)
-    b) delete_index(AO_INDEX)
-    c) delete_index(ARCH_MUR_INDEX)
+def create_test_indices():
 
-    -How to run from terminal:
-    a) python cli.py delete_index case_index
-    b) python cli.py delete_index ao_index
-    c) python cli.py delete_index arch_mur_index
+    INDEX_DICT.update(TEST_INDEX_DICT)
 
-    -How to call task command:
-    a) cf run-task api --command "python cli.py delete_index case_index" -m 2G --name delete_case_index
-    b) cf run-task api --command "python cli.py delete_index ao_index" -m 2G --name delete_ao_index
-    c) cf run-task api --command "python cli.py delete_index arch_mur_index" -m 2G --name delete_arch_mur_index
-    """
-    es_client = create_es_client()
-    logger.info(" Checking if index '{0}' already exist...".format(index_name))
-    if es_client.indices.exists(index=index_name):
-        try:
-            logger.info(" Deleting index '{0}'...".format(index_name))
-            es_client.indices.delete(index_name)
-            # sleep 60 seconds (1 min)
-            time.sleep(60)
-            logger.info(" The index '{0}' is deleted successfully.".format(index_name))
-        except Exception:
-            pass
-    else:
-        logger.error(" The index '{0}' is not found.".format(index_name))
+    for index_name in TEST_INDEX_DICT:
+        create_index(index_name)
+        INDEX_DICT.pop(index_name)
 
 
 def switch_alias(original_index=None, original_alias=None, swapping_index=None):
@@ -621,7 +302,7 @@ def switch_alias(original_index=None, original_alias=None, swapping_index=None):
     -How to call this function in Python code:
     a) switch_alias(index_name, original_alias, swapping_index)
     """
-    original_index = original_index or CASE_INDEX
+    original_index = original_index or constants.CASE_INDEX
     original_alias = original_alias or INDEX_DICT.get(original_index)[1]
     swapping_index = swapping_index or INDEX_DICT.get(original_index)[3]
 
@@ -672,12 +353,12 @@ def restore_from_swapping_index(index_name=None):
     -How to call this function in Python code:
     a) restore_from_swapping_index(index_name)
     """
-    index_name = index_name or CASE_INDEX
+    index_name = index_name or constants.CASE_INDEX
     swapping_index = INDEX_DICT.get(index_name)[3]
     es_client = create_es_client()
 
     # 1) Swith the SEARCH_ALIAS to point to XXXX_SWAP_INDEX instead of index_name(original_index).
-    switch_alias(index_name, SEARCH_ALIAS, swapping_index)
+    switch_alias(index_name, constants.SEARCH_ALIAS, swapping_index)
 
     # 2) Re-create original_index (XXXX_INDEX)
     create_index(index_name)
@@ -687,12 +368,12 @@ def restore_from_swapping_index(index_name=None):
         body={
             "actions": [
                 {"remove": {"index": index_name, "alias": INDEX_DICT.get(index_name)[1]}},
-                {"remove": {"index": index_name, "alias": SEARCH_ALIAS}},
+                {"remove": {"index": index_name, "alias": constants.SEARCH_ALIAS}},
             ]
         }
     )
     logger.info(" Remove aliases '{0}' and '{1}' that point to empty '{2}' successfully.".format(
-        INDEX_DICT.get(index_name)[1], SEARCH_ALIAS, index_name)
+        INDEX_DICT.get(index_name)[1], constants.SEARCH_ALIAS, index_name)
     )
 
     # 4) Re-index XXXX_INDEX based on XXXX_SWAP_INDEX
@@ -728,25 +409,25 @@ def switch_aliases_to_original_index(index_name=None):
     -How to call this function in Python code:
     a) switch_aliases_to_original_index(index_name)
     """
-    index_name = index_name or CASE_INDEX
+    index_name = index_name or constants.CASE_INDEX
     swapping_index = INDEX_DICT.get(index_name)[3]
     es_client = create_es_client()
 
     logger.info(" Moving aliases '{0}' and '{1}' to point to {2}...".format(
-        INDEX_DICT.get(index_name)[1], SEARCH_ALIAS, index_name)
+        INDEX_DICT.get(index_name)[1], constants.SEARCH_ALIAS, index_name)
     )
     es_client.indices.update_aliases(
         body={
             "actions": [
                 {"remove": {"index": swapping_index, "alias": INDEX_DICT.get(index_name)[1]}},
-                {"remove": {"index": swapping_index, "alias": SEARCH_ALIAS}},
+                {"remove": {"index": swapping_index, "alias": constants.SEARCH_ALIAS}},
                 {"add": {"index": index_name, "alias": INDEX_DICT.get(index_name)[1]}},
-                {"add": {"index": index_name, "alias": SEARCH_ALIAS}},
+                {"add": {"index": index_name, "alias": constants.SEARCH_ALIAS}},
             ]
         }
     )
     logger.info(" Moved aliases '{0}' and '{1}' to point to '{2}' successfully.".format(
-        INDEX_DICT.get(index_name)[1], SEARCH_ALIAS, index_name)
+        INDEX_DICT.get(index_name)[1], constants.SEARCH_ALIAS, index_name)
     )
 
     # sleep 60 second (1 min)
@@ -775,12 +456,12 @@ def configure_snapshot_repository(repo_name=None):
     ex3: cf run-task api --command "python cli.py configure_snapshot_repository arch_mur_repo" -m 2G
     --name configure_snapshot_repository_arch_mur
     """
-    repo_name = repo_name or CASE_REPO
+    repo_name = repo_name or constants.CASE_REPO
     es_client = create_es_client()
 
     logger.info(" Configuring snapshot repository: {0}".format(repo_name))
     credentials = get_service_instance_credentials(get_service_instance(
-        S3_PRIVATE_SERVICE_INSTANCE_NAME))
+        constants.S3_PRIVATE_SERVICE_INSTANCE_NAME))
 
     try:
         body = {
@@ -790,7 +471,7 @@ def configure_snapshot_repository(repo_name=None):
                 "region": credentials["region"],
                 "access_key": credentials["access_key_id"],
                 "secret_key": credentials["secret_access_key"],
-                "base_path": S3_BACKUP_DIRECTORY,
+                "base_path": constants.S3_BACKUP_DIRECTORY,
                 "role_arn": env.get_credential("ES_SNAPSHOT_ROLE_ARN"),
             },
         }
@@ -858,7 +539,7 @@ def create_es_snapshot(index_name):
     ex3: cf run-task api --command "python cli.py create_es_snapshot arch_mur_index" -m 2G
     --name create_snapshot_arch_mur
     """
-    index_name = index_name or CASE_INDEX
+    index_name = index_name or constants.CASE_INDEX
     if index_name in INDEX_DICT.keys():
         repo_name = INDEX_DICT.get(index_name)[4]
         prefix_snapshot = INDEX_DICT.get(index_name)[5]
@@ -931,10 +612,10 @@ def restore_es_snapshot(repo_name=None, snapshot_name=None, index_name=None):
     """
     es_client = create_es_client()
 
-    repo_name = repo_name or CASE_REPO
+    repo_name = repo_name or constants.CASE_REPO
     configure_snapshot_repository(repo_name)
 
-    index_name = index_name or CASE_INDEX
+    index_name = index_name or constants.CASE_INDEX
     swapping_index = INDEX_DICT.get(index_name)[3]
     most_recent_snapshot_name = get_most_recent_snapshot(repo_name)
     snapshot_name = snapshot_name or most_recent_snapshot_name
@@ -992,10 +673,10 @@ def restore_es_snapshot_downtime(repo_name=None, snapshot_name=None, index_name=
     """
     es_client = create_es_client()
 
-    repo_name = repo_name or CASE_REPO
+    repo_name = repo_name or constants.CASE_REPO
     configure_snapshot_repository(repo_name)
 
-    index_name = index_name or CASE_INDEX
+    index_name = index_name or constants.CASE_INDEX
 
     if not snapshot_name:
         most_recent_snapshot_name = get_most_recent_snapshot(repo_name)
@@ -1024,7 +705,7 @@ def get_most_recent_snapshot(repo_name=None):
     """
     es_client = create_es_client()
 
-    repo_name = repo_name or CASE_REPO
+    repo_name = repo_name or constants.CASE_REPO
     logger.info(" Retreiving the most recent snapshot...")
     snapshot_list = es_client.snapshot.get(repository=repo_name, snapshot="*").get(
         "snapshots"
@@ -1047,7 +728,7 @@ def display_snapshots(repo_name=None):
     """
     es_client = create_es_client()
 
-    repo_name = repo_name or CASE_REPO
+    repo_name = repo_name or constants.CASE_REPO
     repo_list = [repo_name]
     configure_snapshot_repository(repo_name)
     result = es_client.cat.snapshots(
@@ -1075,7 +756,7 @@ def display_snapshot_detail(repo_name=None, snapshot_name=None):
     -m 2G --name display_snapshot_detail_arch_mur
     """
     es_client = create_es_client()
-    repo_name = repo_name or CASE_REPO
+    repo_name = repo_name or constants.CASE_REPO
     snapshot_name = snapshot_name or "*"
     configure_snapshot_repository(repo_name)
     result = es_client.snapshot.get(
