+++ webservices/tasks/download.py
@@ -15,7 +15,7 @@ from sqlalchemy.dialects import postgresql
 from webservices import utils
 from webservices.common import counts
 from webservices.common.models import db
-from webservices.legal_docs.es_management import S3_BACKUP_DIRECTORY
+from webservices.legal.constants import S3_BACKUP_DIRECTORY
 
 from webservices.tasks import utils as task_utils
 from flask import Flask, current_app
